# docs/swagger.yaml
definitions:
...
  cmd_api-gateway.cancelResp:
    properties:
      status:
        type: string
    type: object
...
  cmd_api-gateway.plansResp:
    properties:
      plans:
        items:
          $ref: '#/definitions/github_com_r0lm0_go-saas-api_internal_billing.Plan'
        type: array
    type: object
...
  cmd_api-gateway.subscribeReq:
    properties:
      email:
        type: string
      plan_slug:
        type: string
    required:
    - email
    - plan_slug
    type: object
  cmd_api-gateway.subscribeResp:
    properties:
      checkout_url:
        type: string
    type: object
  cmd_api-gateway.subscriptionResp:
    properties:
      plan:
        $ref: '#/definitions/github_com_r0lm0_go-saas-api_internal_billing.Plan'
      subscription:
        $ref: '#/definitions/github_com_r0lm0_go-saas-api_internal_billing.Subscription'
    type: object
...
  cmd_api-gateway.webhookResp:
    properties:
      status:
        type: string
    type: object
...
  github_com_r0lm0_go-saas-api_internal_billing.Plan:
    properties:
      amount_cents:
        type: integer
      created_at:
        type: string
      currency:
        type: string
      description:
        type: string
      features:
        additionalProperties: {}
        type: object
      id:
        type: string
      interval:
        type: string
      is_active:
        type: boolean
      max_tokens_per_request:
        type: integer
      messages_per_day:
        type: integer
      name:
        type: string
      slug:
        type: string
      stripe_price_id:
        type: string
      updated_at:
        type: string
    type: object
  github_com_r0lm0_go-saas-api_internal_billing.Subscription:
    properties:
      cancel_at_period_end:
        type: boolean
      canceled_at:
        type: string
      created_at:
        type: string
      current_period_end:
        type: string
      current_period_start:
        type: string
      id:
        type: string
      plan_id:
        type: string
      status:
        type: string
      stripe_customer_id:
        type: string
      stripe_subscription_id:
        type: string
      tenant_id:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/billing/cancel:
    post:
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd_api-gateway.cancelResp'
      security:
      - BearerAuth: []
      summary: Cancel subscription
      tags:
      - billing
  /api/v1/billing/plans:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd_api-gateway.plansResp'
      security:
      - BearerAuth: []
      summary: List billing plans
      tags:
      - billing
  /api/v1/billing/subscribe:
    post:
      consumes:
      - application/json
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: Subscription request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/cmd_api-gateway.subscribeReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd_api-gateway.subscribeResp'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create Stripe checkout session
      tags:
      - billing
  /api/v1/billing/subscription:
    get:
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd_api-gateway.subscriptionResp'
      security:
      - BearerAuth: []
      summary: Get current subscription
      tags:
      - billing
  /api/v1/billing/webhook:
    post:
      consumes:
      - application/json
      parameters:
      - description: Stripe signature
        in: header
        name: Stripe-Signature
        required: true
        type: string
      - description: Webhook payload
        in: body
        name: payload
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd_api-gateway.webhookResp'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Stripe webhook
      tags:
      - billing
...
  /api/v1/stripe/confirm-session:
    post:
      consumes:
      - application/json
      parameters:
      - description: Session ID
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Confirm Stripe checkout session (r3-chat)
      tags:
      - billing
  /api/v1/stripe/create-checkout-session:
    post:
      consumes:
      - application/json
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: Price ID
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create Stripe checkout session (r3-chat)
      tags:
      - billing
  /api/v1/stripe/create-portal-session:
    post:
      consumes:
      - application/json
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create Stripe portal session (r3-chat)
      tags:
      - billing
  /api/v1/stripe/subscription:
    get:
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Stripe subscription (r3-chat)
      tags:
      - billing
  /api/v1/subscriptions:
    post:
      consumes:
      - application/json
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: Plan
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create subscription (r3-chat)
      tags:
      - billing
  /api/v1/subscriptions/cancel:
    post:
      consumes:
      - application/json
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Cancel subscription (r3-chat)
      tags:
      - billing
...
swagger: "2.0"